+++ packages/staker/src/components/TokensPrice.js
@@ -1,12 +1,17 @@
 'use strict';
 
 import React from 'react';
-import {Text} from 'ink';
-import {useTokensPrice} from 'staker-hooks'
+import {View, Text} from 'ink';
+import {useTokensPrice} from 'staker-hooks';
+import _toUpper from 'lodash/toUpper';
 
 export const TokensPrice = ({tokens, fetch}) => {
-  const [tokenPrices] = useTokensPrice(tokens, fetch);
-  return <Text>{JSON.stringify(tokenPrices)}</Text>
+  const [tokenPrices, tokenIdMap] = useTokensPrice(tokens, fetch);
+  return <View>
+    {(Object.keys(tokenPrices)).map(key =>
+      <Text key={key}>{`${(tokenIdMap[key]).toUpperCase()}: ${tokenPrices[key].usd}`}</Text>
+    )}
+  </View>
 }
 
 export default TokensPrice